+++ lib/assets/javascripts/new-dashboard/components/MapCard/CondensedMapCard.vue
@@ -52,7 +52,7 @@
 
     <div class="cell quick-actions" @mouseover="mouseOverChildElement" @mouseleave="mouseOutChildElement">
       <div class="quick-actions__placeholder" v-if="!showInteractiveElements"></div>
-      <MapQuickActions class="quick-actions__element" v-if="showInteractiveElements" :map="visualization" @open="openQuickActions" @close="closeQuickActions" @dataChanged="onDataChanged" :hasShadow="false" />
+      <MapQuickActions class="quick-actions__element" v-if="showInteractiveElements" :map="visualization" @open="openQuickActions" @close="closeQuickActions" @dataChanged="onDataChanged" />
     </div>
   </a>
 </template>